Added helper method for queue_up_anything
This can be used to assert that any job was queued, the behaviour of the
matchers is already tested.

Use case for this helper:

```
expect { thing }.to_not queue_up_anything
```

Previously to write this assertion we would have to write the following:

```
expect { thing }.to_not queue_up(nil)
```

Which is non intuitive, this helper method simply wraps this up into a
readable method.